Restored the functionality, lost in revision 3325, of testing interactive resolution of combat during normal game play.  Specifics:
- Renamed CombatSystem.h's ResolveCombat() to AutoResolveCombat(), and removed
ServerApp::AutoResolveCombat(), which just called ResolveCombat() anyway.  This
is a clearer name for CombatSystem.h's function, and removes the name clash
with the ResolveCombat ServerFSM event, which has been resurrected (see below).

- Collapsed the ServerFSM substates of ProcessingTurn (PreCombatTurnProcessing,
ResolvingCombats, and PostCombatTurnProcessing) into one (ProcessingTurnIdle).
This was necessary, because they were executing code in their ctors that was
causing state transitions to happen.  This leads to undefined behavior that
in our case worked only by accident; adding more state transitions for
resolving individual combats caused everything to blow up.  This was also a
wise change, because these states were vacuous.  They had no event handlers,
except the one that transits to the next substate.  This means that they were
effectively a single state, from an FSM-theoritical perspective, but had been
written as three.

- Resurrected the ResolvingCombat state, which resolves a single combat.  This
state is valuable in that it allows us to handle a single combat at a time,
keeping things more encapsulated (the data for a single combat are stored in
it, instead of data for all combats).

- Pimpled ServerApp::m_fsm.  Now, edits to ServerFSM.h don't cause client
recompiles(!).


git-svn-id: https://freeorion.svn.sourceforge.net/svnroot/freeorion/trunk@3366 dbd5520b-6a0a-0410-a553-f54834df5b05